Moved functions to own .rs file
